chore: remove legacy reranker modules
